refactor: move Popover and add KtPopoverItem
